[partially obsolete] Generic PWM API
The timer implementations expose properties of their available modes,
clock sources etc. through traits class templates.
These are used by the classes of the high-level API
to select appropriate settings for the requested parameters.
This could be extended to different modes of operation.
PWM was just the first use case I tried to implement.

The design has multiple downsides:
- can't handle all the different types of timers without cluttering
  the interface between the generic implementaion and the timer instances
- special cases in the generic code are probably inevitable due to the
  variety of timers even on AVR alone
- relies heavily on template metaprogramming -> difficult to understand
  (C++26 could remove some of the ugliness but it's still over-engineered)

Therefore I decided to discard this approach of a single generic implementation
of the high-level API in favor of individual implementations
for each type of timer which can then use knowledge of the underlying hardware
without the need to squish it through some uniform internal API.

This commit is only here to preserve some of the work I put into this idea
in the unlikely event that I want to return to it.